rationalized segmented document rendering
